src/views/ProjectCardView/ProjectCardView.test.tsx
import mockI18next from '@/mocks/mockI18next';
import axios from 'axios';
import mockProjectCard from '@/mocks/mockProjectCard';
import ProjectCardView from './ProjectCardView';
import { getProjectCardThunk } from '@/reducers/projectCardSlice';
import { renderWithProviders } from '@/utils/testUtils';
import { setupStore } from '@/store';
import { screen, waitFor } from '@testing-library/react';
import { act } from 'react-dom/test-utils';

jest.mock('axios');
jest.mock('react-i18next', () => mockI18next());

const mockedAxios = axios as jest.Mocked<typeof axios>;

describe('ProjectCardView', () => {
  const store = setupStore();

  beforeEach(async () => {
    mockedAxios.get.mockResolvedValue(mockProjectCard);
    await store.dispatch(getProjectCardThunk(mockProjectCard.data.id));
  });

  afterEach(async () => {
    jest.clearAllMocks();
  });

  /**
   * FIXME:
   *
   * Every test here that renders <ProjectCardView /> gets the following error:
   *
   * Warning: An update to ProjectCardHeader inside a test was not wrapped in act(...).
   * When testing, code that causes React state updates should be wrapped into act(...):
   *
   * -----------
   * Tried = using act for every test, using waitFor for every test...
   * -----------
   *
   * The only "solution" that worked (this doesn't seem right...) maybe a bigger problem?:
   *
   *   it.skip('renders the ProjectCardTabs', async () => {
   *     renderWithProviders(<ProjectCardView />);
   *     await act(async () => {
   *        expect(await screen.findByTestId('project-card-tabs-container')).toBeInTheDocument();
   *     })
   *   });
   */
  it.skip('renders the parent container', async () => {
    const { container } = renderWithProviders(<ProjectCardView />);
    expect(container.getElementsByClassName('project-card-container').length).toBe(1);
  });

  it.skip('renders the ProjectCardToolbar', async () => {
    const { container } = renderWithProviders(<ProjectCardView />);
    expect(container.getElementsByClassName('project-card-toolbar-container').length).toBe(1);
  });

  it.skip('renders the ProjectCardHeader', async () => {
    const { container } = renderWithProviders(<ProjectCardView />);
    expect(container.getElementsByClassName('project-card-header-container').length).toBe(1);
  });

  it.skip('renders the ProjectCardTabs', async () => {
    renderWithProviders(<ProjectCardView />);
    await act(async () => {
      expect(await screen.findByTestId('project-card-tabs-container')).toBeInTheDocument();
    });
  });

  it('adds a project card to store if found', async () => {
    const projectCardState = store.getState().projectCard;
    const projectCard = store.getState().projectCard.selectedProjectCard;

    expect(projectCardState).toBeDefined();
    expect(projectCard).not.toBeNull();
    expect(projectCard).toEqual(mockProjectCard.data);
  });
});
